requirements.txt
fastapi
locust
gunicorn
sanic
sanic-redis
aioredis
azure-storage-blob
aiohttp